The iOS app is about 90% completed. Added the ability to view answers. All that is left in the iOS app is to rate the answers.

+++ Code/iOS/Snap2Ask/Snap2Ask/UserModel.h
@@ -11,7 +11,7 @@
 @interface UserModel : NSObject
 
 @property (nonatomic) NSInteger userId;
-@property (nonatomic) NSInteger balance;
+@property (nonatomic) double balance;
 @property (nonatomic) BOOL isTutor;
 @property (nonatomic) BOOL preferredCategoryId;
 @property (strong, nonatomic) NSString *authenticationMode;
